[#936] Extract ICC tag base classes into tags/base.py (item 2)
Move `ICCProfileTag`, `Text`, `TagData`, `DateTimeType`,
`S15Fixed16ArrayType`, `XYZNumber`, `XYZType`, and `ChromaticAdaptionTag`
into a new `DisplayCAL.icc_profile.tags` subpackage (`tags/base.py`).

`LazyLoadTagAODict` moves here too, alongside `ICCProfileTag`, rather than
into `structures.py` as originally sketched in `#936`, since it dispatches
on the tag registries and raises `ICCProfileInvalidError`, neither of which
exist yet. It resolves those via a deferred import, the same pattern used
by `codecs.videoCardGamma()`.

`icc_profile/__init__.py` re-exports everything from `tags.base` so none of
the ~37 importing modules need changes.
