+++ DisplayCAL/icc_profile/tags/__init__.py
@@ -1,0 +1,6 @@
+"""ICC profile tag type classes.
+
+Submodules group tag types by kind (`base`, and further kinds added as the
+DisplayCAL.icc_profile package refactor progresses). This package doesn't
+re-export anything yet, that happens once every tag type has been extracted.
+"""
